Commit before updating quadrature to vectorized version

+++ 2d_galerkin/2d_parabolic_linear/objects/TimeStepping.m
@@ -18,7 +18,7 @@ classdef TimeStepping
 
 			if nargin == 2
 				self.dt    = dt;
-				self.M_t = T / dt;
+				self.M_t = int32(T / dt);
 				self.N_t = self.M_t + 1;
 			end
